+++ src/main/java/com/navangs/maribong/dto/NotificationDTO.java
@@ -4,5 +4,5 @@ public class NotificationDTO {
     String category;
     String userName;
     String message;
-    Integer postId;
+    Long postId;
 }
