+++ src/main/java/com/navangs/maribong/repository/ReplyRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.domain.Limit;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ReplyRepository extends JpaRepository<Reply, Integer> {
-    List<Reply> findByPost_Id(Integer postId);
+    List<Reply> findByPost_Id(Long postId);
 
-    List<Reply> findByPost_Id(Integer postId, Limit limit);
+    List<Reply> findByPost_Id(Long postId, Limit limit);
 }
